Fix [#8] 프로필 수정
- 프로필 이미지를 업데이트할 때, S3 자원을 확보하기 위해 기존 이미지를 삭제하는 로직이 포함되었는데, 변경이 일어나지 않을 때에도, 프론트엔드에서 같은 image key를 전달하여 기존 이미지가 삭제되는 문제가 있어, 일치 여부를 확인하는 과정을 포함했습니다.

+++ src/main/java/cotato/timetile/domain/profile/application/UserProfileService.java
@@ -12,6 +12,7 @@ import cotato.timetile.global.exception.NotFoundException;
 import cotato.timetile.global.exception.UnauthorizedException;
 import cotato.timetile.global.handler.S3Handler;
 import java.util.List;
+import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
@@ -34,9 +35,11 @@ public class UserProfileService {
     @Transactional
     public void update(UserProfileUpdateRequest request, Long userId) {
         User user = userRepository.findById(userId).orElseThrow(UnauthorizedException::failed);
-        s3Handler.deleteNotAllowedFile(request.imageKey());
-        s3Handler.deleteFile(user.getImageKey());
-        user.updateProfile(request.nickname(), request.introduction(), request.imageKey());
+        if (!Objects.equals(request.imageKey(), user.getImageKey())) {
+            s3Handler.deleteNotAllowedFile(request.imageKey());
+            s3Handler.deleteFile(user.getImageKey());
+        }
+        user.updateProfile(request);
         applicationEventPublisher.publishEvent(UserUpdateEvent.of(user));
     }
